Move DefaultProblems to problems project
