+++ platforms/core-execution/workers/src/main/java/org/gradle/workers/internal/DefaultSingleRequestWorkerProcessBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.process.internal.worker;
+package org.gradle.workers.internal;
 
 import org.gradle.api.logging.LogLevel;
 import org.gradle.internal.Cast;
@@ -23,18 +23,23 @@ import org.gradle.internal.logging.events.OutputEventListener;
 import org.gradle.internal.operations.CurrentBuildOperationRef;
 import org.gradle.internal.remote.ObjectConnection;
 import org.gradle.process.internal.JavaExecHandleBuilder;
+import org.gradle.process.internal.worker.DefaultWorkerProcessBuilder;
+import org.gradle.process.internal.worker.RequestHandler;
+import org.gradle.process.internal.worker.SingleRequestWorkerProcessBuilder;
+import org.gradle.process.internal.worker.WorkerProcess;
+import org.gradle.process.internal.worker.WorkerProcessException;
+import org.gradle.process.internal.worker.WorkerProcessSettings;
 import org.gradle.process.internal.worker.request.Receiver;
 import org.gradle.process.internal.worker.request.Request;
 import org.gradle.process.internal.worker.request.RequestArgumentSerializers;
 import org.gradle.process.internal.worker.request.RequestProtocol;
 import org.gradle.process.internal.worker.request.RequestSerializerRegistry;
 import org.gradle.process.internal.worker.request.ResponseProtocol;
-import org.gradle.process.internal.worker.request.WorkerAction;
 
 import java.io.File;
 import java.util.Set;
 
-class DefaultSingleRequestWorkerProcessBuilder<IN, OUT> implements SingleRequestWorkerProcessBuilder<IN, OUT> {
+public class DefaultSingleRequestWorkerProcessBuilder<IN, OUT> implements SingleRequestWorkerProcessBuilder<IN, OUT> {
     private final Class<?> workerImplementation;
     private final DefaultWorkerProcessBuilder builder;
     private final RequestArgumentSerializers argumentSerializers = new RequestArgumentSerializers();
