+++ subprojects/core/src/integTest/groovy/org/gradle/process/internal/AbstractWorkerProcessIntegrationSpec.groovy
@@ -51,12 +51,12 @@ import org.gradle.internal.service.DefaultServiceRegistry
 import org.gradle.internal.service.ServiceRegistryBuilder
 import org.gradle.internal.service.scopes.GlobalScopeServices
 import org.gradle.process.internal.health.memory.MemoryManager
-import org.gradle.process.internal.worker.DefaultWorkerProcessFactory
 import org.gradle.process.internal.worker.child.WorkerProcessClassPathProvider
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.gradle.testfixtures.internal.NativeServicesTestFixture
 import org.gradle.util.GradleVersion
 import org.gradle.util.internal.RedirectStdOutAndErr
+import org.gradle.workers.internal.DefaultWorkerProcessFactory
 import org.junit.Rule
 import spock.lang.Shared
 import spock.lang.Specification
