+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/GradleUserHomeScopeServices.java
@@ -72,7 +72,6 @@ import org.gradle.internal.execution.timeout.TimeoutHandler;
 import org.gradle.internal.execution.timeout.impl.DefaultTimeoutHandler;
 import org.gradle.internal.file.FileAccessTimeJournal;
 import org.gradle.internal.hash.ClassLoaderHierarchyHasher;
-import org.gradle.internal.id.LongIdGenerator;
 import org.gradle.internal.jvm.JavaModuleDetector;
 import org.gradle.internal.jvm.inspection.CachingJvmMetadataDetector;
 import org.gradle.internal.jvm.inspection.DefaultJvmMetadataDetector;
@@ -81,11 +80,8 @@ import org.gradle.internal.jvm.inspection.InvalidInstallationWarningReporter;
 import org.gradle.internal.jvm.inspection.JvmMetadataDetector;
 import org.gradle.internal.jvm.inspection.JvmVersionDetector;
 import org.gradle.internal.jvm.inspection.ReportingJvmMetadataDetector;
-import org.gradle.internal.logging.LoggingManagerInternal;
-import org.gradle.internal.logging.events.OutputEventListener;
 import org.gradle.internal.operations.CurrentBuildOperationRef;
 import org.gradle.internal.reflect.Instantiator;
-import org.gradle.internal.remote.MessagingServer;
 import org.gradle.internal.service.Provides;
 import org.gradle.internal.service.ServiceRegistration;
 import org.gradle.internal.service.ServiceRegistrationProvider;
@@ -93,10 +89,6 @@ import org.gradle.internal.service.ServiceRegistry;
 import org.gradle.internal.vfs.FileSystemAccess;
 import org.gradle.process.internal.ExecFactory;
 import org.gradle.process.internal.ExecHandleFactory;
-import org.gradle.process.internal.JavaExecHandleFactory;
-import org.gradle.process.internal.health.memory.MemoryManager;
-import org.gradle.process.internal.worker.DefaultWorkerProcessFactory;
-import org.gradle.process.internal.worker.WorkerProcessFactory;
 import org.gradle.process.internal.worker.child.WorkerProcessClassPathProvider;
 import org.gradle.util.GradleVersion;
 
@@ -216,32 +208,6 @@ public class GradleUserHomeScopeServices extends WorkerSharedUserHomeScopeServic
             .withObjectFactory(objectFactory)
             .withJavaModuleDetector(javaModuleDetector)
             .build();
-    }
-
-    @Provides
-    WorkerProcessFactory createWorkerProcessFactory(
-        LoggingManagerInternal loggingManagerInternal,
-        MessagingServer messagingServer,
-        ClassPathRegistry classPathRegistry,
-        TemporaryFileProvider temporaryFileProvider,
-        JavaExecHandleFactory execHandleFactory,
-        JvmVersionDetector jvmVersionDetector,
-        MemoryManager memoryManager,
-        GradleUserHomeDirProvider gradleUserHomeDirProvider,
-        OutputEventListener outputEventListener
-    ) {
-        return new DefaultWorkerProcessFactory(
-            loggingManagerInternal,
-            messagingServer,
-            classPathRegistry,
-            new LongIdGenerator(),
-            gradleUserHomeDirProvider.getGradleUserHomeDirectory(),
-            temporaryFileProvider,
-            execHandleFactory,
-            jvmVersionDetector,
-            outputEventListener,
-            memoryManager
-        );
     }
 
     @Provides
