+++ testing/architecture-test/src/changes/archunit-store/internal-api-nullability.txt
@@ -4334,10 +4334,10 @@ Class <org.gradle.process.internal.streams.OutputStreamsForwarder> is not annota
 Class <org.gradle.process.internal.streams.SafeStreams> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (SafeStreams.java:0)
 Class <org.gradle.process.internal.util.LongCommandLineDetectionUtil> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (LongCommandLineDetectionUtil.java:0)
 Class <org.gradle.process.internal.util.MergeOptionsUtil> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (MergeOptionsUtil.java:0)
-Class <org.gradle.process.internal.worker.DefaultMultiRequestWorkerProcessBuilder$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultMultiRequestWorkerProcessBuilder.java:0)
-Class <org.gradle.process.internal.worker.DefaultMultiRequestWorkerProcessBuilder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultMultiRequestWorkerProcessBuilder.java:0)
-Class <org.gradle.process.internal.worker.DefaultSingleRequestWorkerProcessBuilder$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultSingleRequestWorkerProcessBuilder.java:0)
-Class <org.gradle.process.internal.worker.DefaultSingleRequestWorkerProcessBuilder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultSingleRequestWorkerProcessBuilder.java:0)
+Class <org.gradle.workers.internal.DefaultMultiRequestWorkerProcessBuilder$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultMultiRequestWorkerProcessBuilder.java:0)
+Class <org.gradle.workers.internal.DefaultMultiRequestWorkerProcessBuilder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultMultiRequestWorkerProcessBuilder.java:0)
+Class <org.gradle.workers.internal.DefaultSingleRequestWorkerProcessBuilder$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultSingleRequestWorkerProcessBuilder.java:0)
+Class <org.gradle.workers.internal.DefaultSingleRequestWorkerProcessBuilder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultSingleRequestWorkerProcessBuilder.java:0)
 Class <org.gradle.process.internal.worker.DefaultWorkerLoggingProtocol> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultWorkerLoggingProtocol.java:0)
 Class <org.gradle.process.internal.worker.DefaultWorkerProcess$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultWorkerProcess.java:0)
 Class <org.gradle.process.internal.worker.DefaultWorkerProcess$2> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultWorkerProcess.java:0)
@@ -4346,7 +4346,7 @@ Class <org.gradle.process.internal.worker.DefaultWorkerProcessBuilder$MemoryRequ
 Class <org.gradle.process.internal.worker.DefaultWorkerProcessBuilder$WorkerJvmMemoryStatus$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultWorkerProcessBuilder.java:0)
 Class <org.gradle.process.internal.worker.DefaultWorkerProcessBuilder$WorkerJvmMemoryStatus> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultWorkerProcessBuilder.java:0)
 Class <org.gradle.process.internal.worker.DefaultWorkerProcessBuilder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultWorkerProcessBuilder.java:0)
-Class <org.gradle.process.internal.worker.DefaultWorkerProcessFactory> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultWorkerProcessFactory.java:0)
+Class <org.gradle.workers.internal.DefaultWorkerProcessFactory> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (DefaultWorkerProcessFactory.java:0)
 Class <org.gradle.process.internal.worker.GradleWorkerMain> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (GradleWorkerMain.java:0)
 Class <org.gradle.process.internal.worker.MultiRequestClient> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (MultiRequestClient.java:0)
 Class <org.gradle.process.internal.worker.MultiRequestWorkerProcessBuilder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (MultiRequestWorkerProcessBuilder.java:0)
@@ -4391,8 +4391,8 @@ Class <org.gradle.process.internal.worker.request.RequestProtocol> is not annota
 Class <org.gradle.process.internal.worker.request.RequestSerializer> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (RequestSerializer.java:0)
 Class <org.gradle.process.internal.worker.request.RequestSerializerRegistry> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (RequestSerializerRegistry.java:0)
 Class <org.gradle.process.internal.worker.request.ResponseProtocol> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (ResponseProtocol.java:0)
-Class <org.gradle.process.internal.worker.request.WorkerAction$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (WorkerAction.java:0)
-Class <org.gradle.process.internal.worker.request.WorkerAction> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (WorkerAction.java:0)
+Class <org.gradle.workers.internal.WorkerAction$1> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (WorkerAction.java:0)
+Class <org.gradle.workers.internal.WorkerAction> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (WorkerAction.java:0)
 Class <org.gradle.reporting.CodePanelRenderer> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (CodePanelRenderer.java:0)
 Class <org.gradle.reporting.HtmlPageBuilder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (HtmlPageBuilder.java:0)
 Class <org.gradle.reporting.HtmlReportBuilder> is not annotated (directly or via its package) with @org.gradle.api.NonNullApi in (HtmlReportBuilder.java:0)
